backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB driver (async)
motor==3.3.2
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data validation
pydantic[email]==2.5.0
email-validator>=2.1.0

# Environment variables
python-dotenv==1.0.0

